# whatsudow_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatsudow",
        "description": "Aplicación que permite a los organizadores de eventos unificar los datos de sus proveedores de servicios. Con esta app, los organizadores de eventos pueden almacenar y acceder a información importante de los proveedores, comocalendarios de disponibilidad, PDF actualizados, cláusulas y datos de contacto.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/v1/favorites": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Agregar un servicio a favoritos",
                "description": "Agregar un servicio a la lista de favoritos del usuario.",
                "operationId": "36857abd42b1f3ebbf94074ea4a79868",
                "requestBody": {
                    "description": "ID de servicio",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "service_id": {
                                        "type": "integer",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Servicio agregado a favoritos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autenticado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El servicio ya está en favoritos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Eliminar todos los favoritos",
                "description": "Eliminar todos los favoritos del usuario.",
                "operationId": "da943ac7e4ea3c65e3e9703d9037d049",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Todos los favoritos fueron eliminados"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/favorites/{favoriteId}": {
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Eliminar un favorito específico",
                "description": "Eliminar un favorito específico del usuario.",
                "operationId": "ce6925323e260435e62f4833f4dad785",
                "parameters": [
                    {
                        "name": "favoriteId",
                        "in": "path",
                        "description": "ID Favorito",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Favorito eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}